--- operaciones/aprendiz-b.py
@@ -1,14 +1,0 @@
-def division(a, b):
-    if b != 0:
-        return a / b
-    else:
-        return "Error: División entre cero"
-
-def modulo(a, b):
-    if b != 0:
-        return a % b
-    else:
-        return "Error: Módulo entre cero"
-
-def potencia(a, b):
-    return a ** b
