+++ operaciones/aprendiz_b.py
@@ -1,0 +1,11 @@
+def division(a, b):
+    if b != 0:
+        return a / b
+    else:
+        return "Error: División por cero"
+
+def modulo(a, b):
+    return a % b
+
+def potencia(a, b):
+    return a ** b
